Replaced example of `simplify`
Current example of using function `simplify` is `simplify(t + t)`, which gives `2t`, but in Symbolics.jl v. 0.1.10 `t + t` is already `2t` this is not the best example.

Since `2x + 2y` results now in `2x + 2y`, when `simplify(2x + 2y)` gives you `2(x + y)` I think this is better choice for tutorial. In future this can of course change.