Use fixed width text for compound variable names
`capacitor` makes sense as code but does not make as much sense inside of latex output. When you look at it, it italics each of the letters which makes it look like its c times a times ..., which isn't quite right.

Instead, this uses `\texttt{name}` in order to more clearly delineate a value which is a compound name. Using the fixed width text also looks quite "computer-y" so it has a nice connection to `verbose` and is clearly a "computer variable" in the expression.